% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_from_level.R
\name{alias_from_level}
\alias{alias_from_level}
\title{Geography Alias from Level}
\usage{
alias_from_level(query)
}
\arguments{
\item{query}{'Level', the name of a geographic unit as an integer}
}
\value{
A character alias
}
\description{
This function takes the ordinal level of a geographic unit, and returns
its 'alias' as defined in Nucleus.
}
\examples{
\dontrun{
alias_from_level(1)
}
\dontrun{
alias_from_level(c(1, 3))
}

}
