% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_character.R
\name{argument_character}
\alias{argument_character}
\title{Receive Character Argument}
\usage{
argument_character(argument, allow_spaces = FALSE)
}
\arguments{
\item{argument}{Received Python argument}
}
\value{
A reformatted character object
}
\description{
This function processes character arguments received from Python.
}
