% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_logical.R
\name{argument_logical}
\alias{argument_logical}
\title{Receive Logical Argument}
\usage{
argument_logical(argument)
}
\arguments{
\item{argument}{Received Python argument}
}
\value{
A reformatted logical object
}
\description{
This function processes logical arguments received from Python.
}
