% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_numeric.R
\name{argument_numeric}
\alias{argument_numeric}
\title{Receive Numeric Argument}
\usage{
argument_numeric(argument)
}
\arguments{
\item{argument}{Received Python argument}
}
\value{
A reformatted numeric object
}
\description{
This function processes numeric arguments received from Python.
}
