% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_path.R
\name{argument_path}
\alias{argument_path}
\title{Receive Path Argument}
\usage{
argument_path(argument)
}
\arguments{
\item{argument}{Received Python argument}
}
\value{
A reformatted character object describing path
}
\description{
This function processes path arguments received from Python.
}
