% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlap_od_units.R
\name{check_overlap_od_units}
\alias{check_overlap_od_units}
\title{Check OD Table and Unit Geometries Overlap}
\usage{
check_overlap_od_units(od_table, od_ids, units, unit_ids)
}
\arguments{
\item{od_table}{Base origin-destination table}

\item{od_ids}{OD 'start' and 'end' column names as a vector c("start_id", "end_id")}

\item{units}{Dataframe or sf object describing unit geometries}

\item{unit_ids}{Unique ID column name for unit geometries}
}
\value{
A logical output describing whether there are valid overlaps between inputs
}
\description{
This function checks for overlaps between an OD table and unit geometries.
}
