% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_overlap_units_groups.R
\name{check_overlap_units_groups}
\alias{check_overlap_units_groups}
\title{Check Unit Geometries and Groups Overlap}
\usage{
check_overlap_units_groups(units, groups, mode, threshold = 0.1)
}
\arguments{
\item{units}{Dataframe sf object describing unit geometries}

\item{groups}{Dataframe sf object describing groups}
}
\value{
A logical output describing whether there are valid overlaps between inputs
}
\description{
This function checks for overlaps between unit geometries and groups.
}
