% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_children.R
\name{get_children}
\alias{get_children}
\title{Get children geometries}
\usage{
get_children(query, children, year, mode, threshold = 0.1)
}
\arguments{
\item{argument}{Received Python argument}
}
\value{
A reformatted character object
}
\description{
This function pulls a geographic unit's children geometries of a sub-level
}
