% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_distances.R
\name{inter_distances}
\alias{inter_distances}
\title{Unit Inter-Distances}
\usage{
inter_distances(units, id_column = "GEOID", na.rm = FALSE, threshold = NULL)
}
\arguments{
\item{units}{Unit geometries to calculate distances between}

\item{id_column}{Unique identifier column for unit geometries}

\item{na.rm}{A logical determining whether NA values should be omitted from the final output table}

\item{threshold}{Optional upper distance limit, distance values above (non-inclusive) the threshold will be omitted}
}
\value{
A table of inter-unit distances in kilometers, with a 'START' and 'END' column identifying units
}
\description{
This function calculates geodesic distances between given units' centroids.
}
