% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.geographies.R
\docType{data}
\name{k.geographies}
\alias{k.geographies}
\title{Geographic Levels and Aliases}
\format{
A data frame columns:
\describe{
\item{GEOGRAPHY}{Single standard name for geographic unit}
\item{LEVEL}{Ordinal level for geographic unit}
\item{ALIAS}{Alternative names for geographic unit}
\item{ID_LODES}{Name in the package 'lehdr'}
\item{ID_TIGRIS}{Name in the package 'tigris'}
\item{ID_GETACS}{Name in the package 'tidycensus'}
\item{NAMECOLUMN}{ID of name column for tables generated by tidycensus}
}
}
\source{
Created manually for use in Nucleus
}
\usage{
k.geographies
}
\description{
A simple lookup table connecting alias names (like "State") with standard level names.
}
\keyword{datasets}
