% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.states.R
\docType{data}
\name{k.states}
\alias{k.states}
\title{U.S. States and Territories}
\format{
A data frame columns:
\describe{
\item{NAME}{Full name of the area}
\item{SHORT}{The two-leter abbreviation of the area}
\item{GEOID}{The 2-number code of the area}
\item{CONTIGUOUS}{Whether the area is part of the contiguous U.S.}
\item{MAIN}{Whether the area is part of the 50 U.S. states plus D.C.}
\item{X}{Longitude value of area coordinate}
\item{Y}{Latitude value of area coordinate}
\item{AREA_KM2}{Area in square kilometers}
}
}
\source{
Created manually for use in Nucleus
}
\usage{
k.states
}
\description{
A simple lookup table for U.S. states and territories
}
\keyword{datasets}
