% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_from_alias.R
\name{level_from_alias}
\alias{level_from_alias}
\title{Geography Level from Alias}
\usage{
level_from_alias(query)
}
\arguments{
\item{query}{'Alias', the name of a geographic unit as a character}
}
\value{
An integer level
}
\description{
This function takes the 'alias' or name of a geographic unit, and returns
its 'level' as defined in Nucleus.
}
\examples{
\dontrun{
level_from_alias("State")
}
\dontrun{
level_from_alias(c("State", "Tract"))
}

}
