% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_geometry.R
\name{pull_geometry}
\alias{pull_geometry}
\title{Pull Query Geometry}
\usage{
pull_geometry(query, year)
}
\arguments{
\item{query}{A standard Nucleus query}
}
\value{
An sf object
}
\description{
This function pulls a query's geometry from the 'tigris' package.
}
