% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_site.R
\name{visualize_site}
\alias{visualize_site}
\title{Visualize Site}
\usage{
visualize_site(site, device = "jpg", output_directory)
}
\arguments{
\item{site}{A string or integer GEOID describing a geographic area in the U.S.}

\item{device}{File output type as either jpg, png, or pdf.}

\item{output_directory}{A folder to save the final plot.}
}
\value{
A map given argument 'device'
}
\description{
This function produces a minimalistic visualization of a given site.
}
