% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_from_alias.R
\docType{data}
\name{k.geographies}
\alias{k.geographies}
\title{Geography Level from Alias}
\format{
An object of class \code{data.frame} with 15 rows and 7 columns.
}
\usage{
k.geographies
}
\arguments{
\item{query}{'Alias', the name of a geographic unit as a character}
}
\value{
An integer level
}
\description{
This function takes the 'alias' or name of a geographic unit, and returns
its 'level' as defined in Nucleus.
}
\examples{
\dontrun{
level_from_alias("State")
}
\dontrun{
level_from_alias(c("State", "Tract"))
}

Lookup table
}
\keyword{datasets}
