% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_query.R
\name{standardize_query}
\alias{standardize_query}
\title{Standardize Query}
\usage{
standardize_query(query)
}
\arguments{
\item{query}{The character query to standardize}
}
\value{
The standardized query and its level
}
\description{
This function takes an input geographic unit query, and returns a standardized
version of that query in Nucleus' format, or NA if no match exists.
}
